requirements.txt
pandas~=1.1.2
python-telegram-bot~=12.8
schedule~=0.6.0